requirements.txt
python 3.8.15
pip 22.3.1
tensorflow 2.11.0
Pillow 9.4.0
matplotlib 3.6.2
numpy 1.24.1
pandas 1.5.2
opencv-python 4.7.0.68
os 
glob
jupyter 1.0.0
streamlit 1.17.0
keras 2.11.0
imageio 2.24.0